メニューの表示機能をindex.jsに分離
