+++ square-calculation.js
@@ -1,11 +1,7 @@
-import enquirer from "enquirer";
+import { MENU_ID_EASY_MODE, MENU_ID_NORMAL_MODE, main } from "./index.js";
 import Position from "./position.js";
 import Formatter from "./formatter.js";
 import InputCursor from "./input-cursor.js";
-
-const MENU_ID_EASY_MODE = 1;
-const MENU_ID_NORMAL_MODE = 2;
-const MENU_ID_QUIT = 3;
 
 const DIMENTION_OF_EASY_MODE = 5;
 const DIMENTION_OF_NORMAL_MODE = 10;
@@ -18,76 +14,13 @@ class SquareCalculation {
   #matrix;
   #mistakes;
 
-  async execute() {
-    const menuId = await this.#showMenus();
-    try {
-      switch (menuId) {
-        case MENU_ID_EASY_MODE:
-          this.#startNewGame(DIMENTION_OF_EASY_MODE);
-          break;
-        case MENU_ID_NORMAL_MODE:
-          this.#startNewGame(DIMENTION_OF_NORMAL_MODE);
-          break;
-        case MENU_ID_QUIT:
-          process.exit(0);
-          break;
-        default:
-          throw new Error("Unknown menu is selected.");
-          break;
-      }
-    } catch (err) {
-      console.error(err.message);
-      process.exit(1);
+  constructor(menuId) {
+    let dimention;
+    if (menuId === MENU_ID_EASY_MODE) {
+      dimention = DIMENTION_OF_EASY_MODE;
+    } else if (menuId === MENU_ID_NORMAL_MODE) {
+      dimention = DIMENTION_OF_NORMAL_MODE;
     }
-  }
-
-  async #showMenus() {
-    const menus = [
-      {
-        id: MENU_ID_EASY_MODE,
-        name: "Easy Mode (5 x 5 Grid)",
-        description: "Start the game with 25 squares.",
-      },
-      {
-        id: MENU_ID_NORMAL_MODE,
-        name: "Normal Mode (10 x 10 Grid)",
-        description: "Start the game with 100 squares.",
-      },
-      {
-        id: MENU_ID_QUIT,
-        name: "Quit",
-        description: "Exit menu.",
-      },
-    ];
-    const question = {
-      type: "select",
-      name: "value",
-      message: "Please select a menu.",
-      choices: menus.map((menu) => ({
-        name: menu.name,
-        value: menu.id,
-      })),
-      result() {
-        return this.focused.value;
-      },
-      footer() {
-        const descriptions = menus.map((mode) => mode.description);
-        return "\n  " + descriptions[this.index];
-      },
-    };
-
-    this.#updateDisplay();
-    let answer;
-    try {
-      answer = await enquirer.prompt(question);
-    } catch (err) {
-      console.error("An unexpected error has occurred.");
-      process.exit(1);
-    }
-    return answer.value;
-  }
-
-  #startNewGame(dimention) {
     this.#dimention = dimention;
     this.#position = new Position(dimention);
     this.#formatter = new Formatter(dimention);
@@ -95,6 +28,24 @@ class SquareCalculation {
     this.#initMatrix();
     this.#updateDisplay();
     this.#inputCursor.startKeyInput(this.update.bind(this));
+  }
+
+  update(numStr) {
+    const num = Number.parseInt(numStr);
+    if (isNaN(num)) {
+      return;
+    }
+
+    this.#matrix[this.#position.row()][this.#position.column()] = num;
+    this.#position.update();
+    this.#inputCursor.updateCursorPosition();
+
+    if (this.#hasGameEnded()) {
+      this.#inputCursor.stopKeyInput();
+      this.#checkAnswer();
+      this.#updateDisplay();
+      this.#waitForSomeKeyInput();
+    }
   }
 
   #initMatrix() {
@@ -127,24 +78,6 @@ class SquareCalculation {
     }
   }
 
-  update(numStr) {
-    const num = Number.parseInt(numStr);
-    if (isNaN(num)) {
-      return;
-    }
-
-    this.#matrix[this.#position.row()][this.#position.column()] = num;
-    this.#position.update();
-    this.#inputCursor.updateCursorPosition();
-
-    if (this.#hasGameEnded()) {
-      this.#inputCursor.stopKeyInput();
-      this.#checkAnswer();
-      this.#updateDisplay();
-      this.#waitForSomeKeyInput();
-    }
-  }
-
   #hasGameEnded() {
     return this.#position.row() > this.#dimention;
   }
@@ -167,18 +100,8 @@ class SquareCalculation {
     process.stdin.once("data", (_key) => {
       process.stdin.pause();
       process.stdin.removeAllListeners("data");
-      this.#releaseAllResources();
-      this.execute();
+      main();
     });
-  }
-
-  #releaseAllResources() {
-    this.#dimention = null;
-    this.#position = null;
-    this.#inputCursor = null;
-    this.#formatter = null;
-    this.#matrix = null;
-    this.#mistakes = null;
   }
 }
